requirements.txt
ultralytics==8.0.145
pillow==10.0.0
fastapi==0.99.1
python-multipart==0.0.6
uvicorn==0.22.0
numpy>=1.22.2
torch>=2.0.0